+++ PoliNetworkBot_CSharp/Code/Objects/InfoBot/BotDisguisedAsUserBotInfo.cs
@@ -26,7 +26,7 @@ namespace PoliNetworkBot_CSharp.Code.Objects.InfoBot
         {
             try
             {
-                return Convert.ToInt32(KeyValuePairs[ConstConfigBot.UserId]);
+                return Convert.ToInt64(KeyValuePairs[ConstConfigBot.UserId]);
             }
             catch
             {
@@ -38,7 +38,7 @@ namespace PoliNetworkBot_CSharp.Code.Objects.InfoBot
         {
             try
             {
-                return Convert.ToInt32(KeyValuePairs[ConstConfigBot.ApiId]);
+                return Convert.ToInt64(KeyValuePairs[ConstConfigBot.ApiId]);
             }
             catch
             {
@@ -68,7 +68,7 @@ namespace PoliNetworkBot_CSharp.Code.Objects.InfoBot
         {
             try
             {
-                KeyValuePairs[ConstConfigBot.ApiId] = Convert.ToInt32(v);
+                KeyValuePairs[ConstConfigBot.ApiId] = Convert.ToInt64(v);
                 return true;
             }
             catch
@@ -95,7 +95,7 @@ namespace PoliNetworkBot_CSharp.Code.Objects.InfoBot
         {
             try
             {
-                KeyValuePairs[ConstConfigBot.UserId] = Convert.ToInt32(v);
+                KeyValuePairs[ConstConfigBot.UserId] = Convert.ToInt64(v);
                 return true;
             }
             catch
